Add sort for topic
